\name{getAllSites}
\alias{getAllSites}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get all site numbers.
}
\description{
This function simply retrieves all the site numbers from the server.
}
\usage{
getAllSites()
}

\details{
This function retrieves all site numbers within the database off the server.
}
\value{
A data frame containing all site numbers (roughly 3000).
}
\author{
Jason Lessels <jlessels@gmail.com>
}

